refactor: rename `common` to `types`
